login-offline: adapt androidTests
