--- core/src/androidTest/java/org/hisp/dhis/android/core/LogoutCallRealIntegrationShould.java
@@ -1,118 +1,0 @@
-package org.hisp.dhis.android.core;
-
-import org.hisp.dhis.android.core.common.D2Factory;
-import org.hisp.dhis.android.core.common.EventCallFactory;
-import org.hisp.dhis.android.core.data.database.AbsStoreTestCase;
-import org.hisp.dhis.android.core.data.database.DatabaseAssert;
-import org.hisp.dhis.android.core.data.server.RealServerMother;
-import org.hisp.dhis.android.core.event.EventEndpointCall;
-import org.hisp.dhis.android.core.event.EventModel;
-import org.hisp.dhis.android.core.organisationunit.OrganisationUnitModel;
-import org.hisp.dhis.android.core.program.ProgramModel;
-import org.hisp.dhis.android.core.resource.ResourceModel;
-import org.hisp.dhis.android.core.user.AuthenticatedUserModel;
-import org.hisp.dhis.android.core.user.UserModel;
-import org.junit.Before;
-
-import java.io.IOException;
-
-import static com.google.common.truth.Truth.assertThat;
-import static org.hisp.dhis.android.core.data.database.SqliteCheckerUtility.isDatabaseEmpty;
-import static org.hisp.dhis.android.core.data.database.SqliteCheckerUtility.isTableEmpty;
-
-public class LogoutCallRealIntegrationShould extends AbsStoreTestCase {
-    private D2 d2;
-
-    @Before
-    @Override
-    public void setUp() throws IOException {
-        super.setUp();
-
-        d2 = D2Factory.create(RealServerMother.url, databaseAdapter());
-    }
-
-    //@Test
-    public void have_empty_database_when_wipe_db_after_sync_metadata() throws Exception {
-        d2.logIn("android", "Android123").call();
-
-        d2.syncMetaData().call();
-
-        DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
-
-        d2.wipeDB().call();
-
-        DatabaseAssert.assertThatDatabase(databaseAdapter()).isEmpty();
-    }
-
-    //@Test
-    public void have_empty_database_when_wipe_db_after_sync_data() throws Exception {
-        d2.logIn("android", "Android123").call();
-
-        d2.syncMetaData().call();
-
-        EventEndpointCall eventCall = EventCallFactory.create(
-                d2.retrofit(), databaseAdapter(), "DiszpKrYNg8", 0);
-
-        eventCall.call();
-
-        DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
-
-        d2.wipeDB().call();
-
-        DatabaseAssert.assertThatDatabase(databaseAdapter()).isEmpty();
-    }
-
-    //@Test
-    public void delete_autenticate_user_table_only_when_log_out_after_sync_data() throws Exception {
-        d2.logIn("android", "Android123").call();
-
-        d2.syncMetaData().call();
-
-        EventEndpointCall eventCall = EventCallFactory.create(
-                d2.retrofit(), databaseAdapter(), "DiszpKrYNg8", 0);
-
-        eventCall.call();
-
-        assertThat(isDatabaseEmpty(databaseAdapter())).isFalse();
-
-        d2.logout().call();
-
-        assertThat(isDatabaseEmpty(databaseAdapter())).isFalse();
-        assertThat(isTableEmpty(databaseAdapter(), EventModel.TABLE)).isFalse();
-
-        assertThat(isTableEmpty(databaseAdapter(), AuthenticatedUserModel.TABLE)).isTrue();
-    }
-
-    //@Test
-    public void delete_autenticate_user_table_only_when_log_out_after_sync_metadata()
-            throws Exception {
-        d2.logIn("android", "Android123").call();
-
-        d2.syncMetaData().call();
-
-        assertThat(isDatabaseEmpty(databaseAdapter())).isFalse();
-
-        d2.logout().call();
-
-        assertThat(isDatabaseEmpty(databaseAdapter())).isFalse();
-
-        assertThat(isTableEmpty(databaseAdapter(), EventModel.TABLE)).isTrue();
-
-        assertThat(isTableEmpty(databaseAdapter(), UserModel.TABLE)).isFalse();
-
-        assertThat(isTableEmpty(databaseAdapter(), OrganisationUnitModel.TABLE)).isFalse();
-
-        assertThat(isTableEmpty(databaseAdapter(), ProgramModel.TABLE)).isFalse();
-
-        assertThat(isTableEmpty(databaseAdapter(), ResourceModel.TABLE)).isFalse();
-
-        assertThat(isTableEmpty(databaseAdapter(), AuthenticatedUserModel.TABLE)).isTrue();
-    }
-
-    //@Test
-    public void response_successful_on_login_logout_and_login() throws Exception {
-        d2.logIn("android", "Android123").call();
-        d2.logout().call();
-        d2.logIn("android", "Android123").call();
-    }
-}
