+++ core/src/androidTest/java/org/hisp/dhis/android/core/calls/WipeDBCallMockIntegrationShould.java
@@ -1,0 +1,96 @@
+package org.hisp.dhis.android.core.calls;
+
+import org.hisp.dhis.android.core.D2;
+import org.hisp.dhis.android.core.common.D2Factory;
+import org.hisp.dhis.android.core.common.EventCallFactory;
+import org.hisp.dhis.android.core.data.database.AbsStoreTestCase;
+import org.hisp.dhis.android.core.data.database.DatabaseAssert;
+import org.hisp.dhis.android.core.data.file.AssetsFileReader;
+import org.hisp.dhis.android.core.data.server.Dhis2MockServer;
+import org.hisp.dhis.android.core.event.Event;
+import org.hisp.dhis.android.core.event.EventEndpointCall;
+import org.hisp.dhis.android.core.event.EventPersistenceCall;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import java.io.IOException;
+import java.util.List;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+
+public class WipeDBCallMockIntegrationShould extends AbsStoreTestCase {
+
+    private Dhis2MockServer dhis2MockServer;
+    private D2 d2;
+
+    @Override
+    @Before
+    public void setUp() throws IOException {
+        super.setUp();
+
+        dhis2MockServer = new Dhis2MockServer(new AssetsFileReader());
+
+        d2 = D2Factory.create(dhis2MockServer.getBaseEndpoint(), databaseAdapter());
+    }
+
+    @Override
+    @After
+    public void tearDown() throws IOException {
+        super.tearDown();
+
+        dhis2MockServer.shutdown();
+    }
+
+    @Test
+    public void have_empty_database_when_wipe_db_after_sync_meta_data() throws Exception {
+        givenALoginInDatabase();
+
+        givenAMetadataInDatabase();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
+
+        d2.wipeDB().call();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isEmpty();
+    }
+
+    @Test
+    public void have_empty_database_when_wipe_db_after_sync_data() throws Exception {
+        givenALoginInDatabase();
+
+        givenAMetadataInDatabase();
+
+        givenAEventInDatabase();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
+
+        d2.wipeDB().call();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isEmpty();
+    }
+
+    private void givenALoginInDatabase() throws Exception {
+        dhis2MockServer.enqueueLoginResponses();
+        d2.logIn("user", "password").call();
+    }
+
+    private void givenAMetadataInDatabase() throws Exception {
+        dhis2MockServer.enqueueMetadataResponses();
+        d2.syncMetaData().call();
+    }
+
+    private void givenAEventInDatabase() throws Exception {
+        EventEndpointCall eventCall = EventCallFactory.create(
+                d2.retrofit(), databaseAdapter(), "DiszpKrYNg8", 0);
+
+        dhis2MockServer.enqueueMockResponse("events.json");
+
+        List<Event> events = eventCall.call();
+
+        EventPersistenceCall.create(databaseAdapter(), d2.retrofit(), events).call();
+
+        assertThat(events.isEmpty(), is(false));
+    }
+}
