+++ core/src/androidTest/java/org/hisp/dhis/android/core/calls/WipeDBCallRealIntegrationShould.java
@@ -1,0 +1,55 @@
+package org.hisp.dhis.android.core.calls;
+
+import org.hisp.dhis.android.core.D2;
+import org.hisp.dhis.android.core.common.D2Factory;
+import org.hisp.dhis.android.core.common.EventCallFactory;
+import org.hisp.dhis.android.core.data.database.AbsStoreTestCase;
+import org.hisp.dhis.android.core.data.database.DatabaseAssert;
+import org.hisp.dhis.android.core.data.server.RealServerMother;
+import org.hisp.dhis.android.core.event.EventEndpointCall;
+import org.junit.Before;
+
+import java.io.IOException;
+
+public class WipeDBCallRealIntegrationShould extends AbsStoreTestCase {
+    private D2 d2;
+
+    @Before
+    @Override
+    public void setUp() throws IOException {
+        super.setUp();
+
+        d2 = D2Factory.create(RealServerMother.url, databaseAdapter());
+    }
+
+    //@Test
+    public void have_empty_database_when_wipe_db_after_sync_metadata() throws Exception {
+        d2.logIn("android", "Android123").call();
+
+        d2.syncMetaData().call();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
+
+        d2.wipeDB().call();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isEmpty();
+    }
+
+    //@Test
+    public void have_empty_database_when_wipe_db_after_sync_data() throws Exception {
+        d2.logIn("android", "Android123").call();
+
+        d2.syncMetaData().call();
+
+        EventEndpointCall eventCall = EventCallFactory.create(
+                d2.retrofit(), databaseAdapter(), "DiszpKrYNg8", 0);
+
+        eventCall.call();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isNotEmpty();
+
+        d2.wipeDB().call();
+
+        DatabaseAssert.assertThatDatabase(databaseAdapter()).isEmpty();
+    }
+}
